+++ lib/altive_chat_room.dart
@@ -1,3 +1,4 @@
 export 'src/altive_chat_room.dart';
+export 'src/chat_deletion_confirmation.dart';
 export 'src/chat_message_card.dart';
 export 'src/models.dart';